pyproject.toml
[project]
name = "try-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.4.1",
    "openai>=2.15.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]


[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["rag", "data"]
